Validaciones

+++ src/pages/DashboardCiudadesPage.js
@@ -2,18 +2,18 @@ import { map } from "@firebase/util";
 import React from "react";
 import CiudadesLista from "../components/CiudadesLista/CiudadesLista";
 import '../css/dashboardCiudades.css'
-import {queryCiudades, deleteCiudad, createCiudad} from "../services/ciudades"
+import { queryCiudades, deleteCiudad, createCiudad } from "../services/ciudades"
 import { useNavigate } from 'react-router-dom';
+import swal from "sweetalert";
 
 
-function DashboardCiudadesPage(){
+function DashboardCiudadesPage() {
 
     const [ciudades, setCiudades] = React.useState([])
     const [loading, setLoading] = React.useState(false)
     const navigate = useNavigate()
 
     React.useEffect(() => {
-
         setLoading(true)
         queryCiudades().then((response) => {
             setCiudades(response)
@@ -22,14 +22,33 @@ function DashboardCiudadesPage(){
 
     }, [])
 
-    
 
-    const handleDelete = (ciudadId) =>{
-        setLoading(true)
-        deleteCiudad(ciudadId).then(() => {
-            setCiudades(ciudades.filter( ciudad => ciudadId != ciudad.id))
-            setLoading(false)
+
+    const handleDelete = (ciudadId) => {
+        swal({
+            title: "¿Estás seguro?",
+            text: "Una vez eliminado, no se podrá recuperar los datos almacenados",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
         })
+            .then((willDelete) => {
+                if (willDelete) {
+                    swal("Se eliminó con exito.", {
+                        icon: "success",
+
+                    }
+
+                    );
+                    setLoading(true)
+                    deleteCiudad(ciudadId).then(() => {
+                        setCiudades(ciudades.filter(ciudad => ciudadId != ciudad.id))
+                        setLoading(false)
+                    })
+                } else {
+                    swal("Your imaginary file is safe!");
+                }
+            });
     }
 
 
@@ -50,29 +69,29 @@ function DashboardCiudadesPage(){
 
     return (
         <>
-        <div className="container">
+            <div className="container">
 
-            {loading ? <div>cargando</div> : null}
-            {ciudades.map(({ id, nombre }) => (
-                <>
-                    <div className="data">
-                        <div className="ciudades">{nombre}</div>
-                        <div className="botonesContainer">
-                            <button onClick={() => navigate(`/dashboardCiudades/${id}`)}> Editar </button>
-                            <button onClick={(() => {handleDelete(id)})}> Eliminar </button>
+                {loading ? <div>cargando</div> : null}
+                {ciudades.map(({ id, nombre }) => (
+                    <>
+                        <div className="data">
+                            <div className="ciudades">{nombre}</div>
+                            <div className="botonesContainer">
+                                <button onClick={() => navigate(`/dashboardCiudades/${id}`)}> Editar </button>
+                                <button onClick={(() => { handleDelete(id) })}> Eliminar </button>
+                            </div>
                         </div>
-                    </div>
-                </>
-            ))}
-            <div className="title">Agregar una nueva Ciudad AQUI</div>
-            <form onSubmit={handleSubmit}>
-            
-            <button onClick = {() => navigate(`/dashboardCiudades/create`)}>Submit</button>
-        </form> 
-        <div>Aqui para Devolverse al Dashboard</div>
-        <button onClick={() => navigate(`/dashboard`)}>Click!!!</button>
-        </div>
-    </>
+                    </>
+                ))}
+                <div className="title">Agregar una nueva Ciudad AQUI</div>
+                <form onSubmit={handleSubmit}>
+
+                    <button onClick={() => navigate(`/dashboardCiudades/create`)}>Submit</button>
+                </form>
+                <div>Aqui para Devolverse al Dashboard</div>
+                <button onClick={() => navigate(`/dashboard`)}>Click!!!</button>
+            </div>
+        </>
     );
 
 }
