+++ src/pages/DashboardHotelesPage.js
@@ -1,11 +1,11 @@
 import React from "react";
 import { map } from "@firebase/util";
 import '../css/dashboardHoteles.css'
-import {queryHoteles, deleteHotel} from "../services/hoteles"
+import { queryHoteles, deleteHotel } from "../services/hoteles"
 import { useNavigate } from 'react-router-dom';
+import swal from "sweetalert";
 
-
-function DashboardHotelesPage(){
+function DashboardHotelesPage() {
 
     const [hoteles, setHoteles] = React.useState([])
     const [loading, setLoading] = React.useState(false)
@@ -21,12 +21,31 @@ function DashboardHotelesPage(){
 
     }, [])
 
-    const handleDelete = (hotelId) =>{
-        setLoading(true)
-        deleteHotel(hotelId).then(() => {
-            setHoteles(hoteles.filter( hotel => hotelId != hotel.id))
-            setLoading(false)
+    const handleDelete = (hotelId) => {
+        swal({
+            title: "¿Estás seguro?",
+            text: "Una vez eliminado, no se podrá recuperar los datos almacenados",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
         })
+            .then((willDelete) => {
+                if (willDelete) {
+                    swal("Se eliminó con exito.", {
+                        icon: "success",
+
+                    }
+
+                    );
+                    setLoading(true)
+                    deleteHotel(hotelId).then(() => {
+                        setHoteles(hoteles.filter(hotel => hotelId != hotel.id))
+                        setLoading(false)
+                    })
+                } else {
+                    swal("Your imaginary file is safe!");
+                }
+            });
     }
 
     return (
@@ -38,18 +57,18 @@ function DashboardHotelesPage(){
                         <div className="hoteles">{nombre}</div>
                         <div className="botonesContainer">
                             <button onClick={() => navigate(`/dashboardHoteles/${id}`)}> Editar </button>
-                            <button onClick={(() => {handleDelete(id)})}> Eliminar </button>
+                            <button onClick={(() => { handleDelete(id) })}> Eliminar </button>
                         </div>
                     </div>
                 </>
             ))}
-        <div>Aqui para agregar Nuevo Hotel</div>
-        <button onClick={() => navigate(`/dashboardHoteles/create`)}>Click!!!</button>
-        
-        <div>Aqui para Devolverse al Dashboard</div>
-        <button className="BotonesNormales" onClick={() => navigate(`/dashboard`)}>Click!!!</button>
-        <div>Aqui para Modificar las habitaciones de tus Hoteles</div>
-        <button className="BotonesNormales" onClick={() => navigate(`/dashboardHabitaciones`)}>Habitaciones</button>
+            <div>Aqui para agregar Nuevo Hotel</div>
+            <button onClick={() => navigate(`/dashboardHoteles/create`)}>Click!!!</button>
+
+            <div>Aqui para Devolverse al Dashboard</div>
+            <button className="BotonesNormales" onClick={() => navigate(`/dashboard`)}>Click!!!</button>
+            <div>Aqui para Modificar las habitaciones de tus Hoteles</div>
+            <button className="BotonesNormales" onClick={() => navigate(`/dashboardHabitaciones`)}>Habitaciones</button>
         </div>
     );
 }
